Store all operators, combinators, and macros in a single library
This commit introduces a new interface Function, which is extended by
the Operator, Combinator, and Macro interfaces. This allows the
interpreter to call each one in the same way with an 'invoke' method,
leaving it to the implementation to use the parameters as needed.

The added benefit is that there doesn't need to be separate libraries
for each type of function. A single library effectually enforces a
single namespace.