Add a "joy" runtime, and support "--mode" for choosing runtimes
This commit refactors the Main class quite a bit. It adds support for
command-specific options. Specifically, the "--mode" option for both
eval and repl. The eval command now requires a "--" before evaluating
a program if options are provided. The repl command now has an
"--eval" option that will evaulate a program immediately after
starting the REPL.