Implement line comments for Joy and Rejoice
Joy line comments begin with '#'. Rejoice line comments begin with
'//'. Rejoice has different line comments in order to save '#' for
other uses.

In order to implement these differently for each runtime without
writing separate lexers, this commit adds a new interface called
'LexerRule'. Each rule contains a character to dispatch on, and a
mini-lexer for how to handle the input stream. Like 'Lexer', the rule
accepts a 'PushbackReader' and returns a 'Lexer.Token'.

For Joy, the implementation is fairly basic. Once a '#" is found, all
the characters until there's a new line ('\r', '\n', or '\r\n') are
read and returned as a 'Comment' token. The parser ultimately drops
it, however, it may do something with it in the future.

For Rejoice, the implementation is a little more involved. The symbol
`/` is a valid token on its own, so that is supported. So is any
symbol that begins or ends with a `/`. A symbol cannot have `/` in the
middle of it, because that is reserved for future use. If `//` is
found, then the behavior is the same as the Joy runtime.