Update Java class naming for functions with special characters
Each Java class name has a single letter prefix.
- O: operator
- C: combinator
- M: macro

The rest of the Java class name represnts the name of the function.

If the name doesn't begin with an underscore, then the text represents
the literal name of the function for Rejoice. (Joy names may differ,
but will use the same Rejoice classes.)

If the name does begin with an underscore, then the following text is
title-cased and is a representation of the function in English.. This
is because the symbol for the function is not a valid Java clas sname.

If the name contains the characters `_E_` then that means the name has
an exclamation mark. There's also `_Q_` for a question mark.