Rewrite the atom stream with macros instead of aggregating
This commit replaces the compiler with a rewriter, and changes the
semantics of macros. Before, macros would generate non-atomic values
and place them on the stream for the interpreter to process. Now,
macros only handle atomic values and rewrite the stream, leaving
functions to generate non-atomic values.

As a consequence, this introduced the `list` and `define!` functions,
which generate lists and definitions, respectively, by looking for
values on the stack.