Implement basic modules
The module system was inspired by Joy and OCaml. The current
implementation allows for functions to be grouped into a module.

Modules can either be included and required. Included modules have
their functions automatically exported by the module including it.
Required modules allow for a module to be references without the
module name within the module requring it, but the functions are not
exported.

The default module is 'core' and is automatically required by every
module. The core module includes an internal 'native' module which has
all the functions that are implemented in Java.