Remove the Joy implementation
I've decided to remove the Joy implementation for now. While it was
fun at first to build both Joy and Rejoice simultaneously, which as
much shared code as possible, it has become too much to do at once.
I've also come to the conclusion that Joy could be implemented in
terms of Rejoice as the language evolves. This would at least require
`define-syntax` and a way to customize the lexical analysis phase.