Make evaluation incremental, again
Instead of each phase of evaluation being complete before the next
phase begins, each phase calls into the next phase only when it needs
the next thing to evaulate. This adds a bit of complexity, however it
will allow macros to be scoped to modules, as they can be defined and
then subsequently used. It also may allow functions to be used to
influence parsing.

This also:
- Adds a comment atom which the interpreter ignores.
- Fixes a bug where 'require' was calling '%include'.
- Makes %list a function that can evaluate itself, which means
  interpreted functions don't need a special wrapper, or the
  interpreter to be available from the context. (After all, functions
  are lists of values, while the interpreter accepts a list of atoms.)
- Moves macros to the context.