Unifty functions and macros as invocables
Functions and macros need to exist in the same namespace. In order to
keep the management of the namespaces simple and typed, they both
share a parent type: invocable.

A function accepts an environment and a stack. A macro accepts an
environment and an input stream. An invocable accepts an environment,
a stack, and an input stream. The interfaces for functions and macros
have not changed, instead they are specializations over invocable.

Because of this, the interpreter does not need to be aware of macros.
Macros can now be self-sufficent and interpret their rewritten input
stream since they are an invocable and have the stack available.