Macros generate values from the input stream.
While macros still behave the same way, in that they pull values from
the input stream and use them to generate new values, they are not
limited to rewriting atoms (a subset of values). For example,
processing a stack, the program `[ 1 2 3 ]` is rewritten to a stack
value, rather than a program that creates the same stack value.